Introduces SubscriptionsPage component
